Fix issue when importing utils from getTinyMceInitOptions, move validation functions to fieldPropertyUtils and update imports
